+++ src/azure-firewall/azext_firewall/commands.py
@@ -70,10 +70,10 @@ def load_command_table(self, _):
     # region AzureFirewalls
     with self.command_group('network firewall', network_firewall_sdk) as g:
         g.custom_command('create', 'create_azure_firewall')
-        g.command('delete', 'delete')
+        g.command('delete', 'begin_delete')
         g.custom_command('list', 'list_azure_firewalls')
         g.show_command('show')
-        g.generic_update_command('update', custom_func_name='update_azure_firewall')
+        g.generic_update_command('update', setter_name="begin_create_or_update", custom_func_name='update_azure_firewall')
 
     with self.command_group('network firewall', network_firewall_sdk_start, client_factory=cf_firewalls_2020_11_01) as g:
         g.custom_command('start', 'start_azure_firewall', is_preview=True, validator=validate_af_start)
@@ -83,7 +83,7 @@ def load_command_table(self, _):
         g.custom_command('create', 'create_azure_firewall_threat_intel_allowlist')
         g.custom_command('delete', 'delete_azure_firewall_threat_intel_allowlist')
         g.custom_show_command('show', 'show_azure_firewall_threat_intel_allowlist')
-        g.generic_update_command('update', custom_func_name='update_azure_firewall_threat_intel_allowlist')
+        g.generic_update_command('update', setter_name='begin_create_or_update', custom_func_name='update_azure_firewall_threat_intel_allowlist')
 
     with self.command_group('network firewall ip-config', network_util) as g:
         g.custom_command('create', 'create_af_ip_configuration')
@@ -134,7 +134,7 @@ def load_command_table(self, _):
     # region AzureFirewallPolicies
     with self.command_group('network firewall policy', network_firewall_policies_sdk, resource_type=CUSTOM_FIREWALL, min_api='2019-07-01') as g:
         g.custom_command('create', 'create_azure_firewall_policies', exception_handler=exception_handler)
-        g.command('delete', 'delete')
+        g.command('delete', 'begin_delete')
         g.custom_command('list', 'list_azure_firewall_policies')
         g.show_command('show')
         g.generic_update_command('update', custom_func_name='update_azure_firewall_policies',
@@ -149,8 +149,8 @@ def load_command_table(self, _):
 
     with self.command_group('network firewall policy rule-collection-group', network_firewall_policy_rule_groups, resource_type=CUSTOM_FIREWALL, is_preview=True) as g:
         g.custom_command('create', 'create_azure_firewall_policy_rule_collection_group', exception_handler=exception_handler)
-        g.generic_update_command('update', custom_func_name='update_azure_firewall_policy_rule_collection_group', exception_handler=exception_handler)
-        g.command('delete', 'delete')
+        g.generic_update_command('update', custom_func_name='update_azure_firewall_policy_rule_collection_group', setter_name='begin_create_or_update', exception_handler=exception_handler)
+        g.command('delete', 'begin_delete')
         g.show_command('show')
         g.command('list', 'list')
 
@@ -162,6 +162,6 @@ def load_command_table(self, _):
 
     with self.command_group('network firewall policy rule-collection-group collection rule', network_firewall_policy_rule_groups, resource_type=CUSTOM_FIREWALL, is_preview=True) as g:
         g.custom_command('add', 'add_azure_firewall_policy_filter_rule', exception_handler=exception_handler)
-        g.generic_update_command('update', custom_func_name='update_azure_firewall_policy_filter_rule', exception_handler=exception_handler)
+        g.generic_update_command('update', setter_name='begin_create_or_update', custom_func_name='update_azure_firewall_policy_filter_rule', exception_handler=exception_handler)
         g.custom_command('remove', 'remove_azure_firewall_policy_filter_rule')
     # endregion