+++ src/blueprint/azext_blueprint/tests/latest/test_blueprint_scenario.py
@@ -5,6 +5,10 @@
 
 import os
 import unittest
+import json
+import filecmp
+from pathlib import Path
+import shutil
 
 from azure_devtools.scenario_tests import AllowLargeResponse
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer,
@@ -288,3 +292,90 @@ class BlueprintScenarioTest(ScenarioTest):
                  '--subscription "{subscription}" '
                  '--name "storage-rg" '
                  '-y')
+
+    @AllowLargeResponse()
+    @ResourceGroupPreparer(name_prefix='cli_test_blueprint_export')
+    def test_blueprint_export(self, resource_group):
+        # same procedure as testing the import, build a blueprint to export
+        self.kwargs.update({
+            'blueprintName': self.create_random_name(prefix='test-imported-bp', length=24),
+            'subscription': self.get_subscription_id(),
+        })
+
+        # this will overwrite the previous settings
+        self.cmd(
+            'az blueprint import '
+            '--name "{blueprintName}" '
+            '--input-path "src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/input" '
+            '-y',
+            checks=[])
+
+        self.cmd(
+            'az blueprint export '
+            '--output-path "src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/exported" '
+            '--name "{blueprintName}" '
+            '--yes',
+            checks=[])
+
+        # check if the import and output artifacts are equal in content
+        input_blueprint = "src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/input/blueprint.json"
+        input_artifact_directory = "src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/input/artifacts"
+        output_blueprint = f"src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/exported/{self.kwargs['blueprintName']}/blueprint.json"
+        output_artifact_directory = f"src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/exported/{self.kwargs['blueprintName']}/artifacts"
+        output_path = Path(f"src/blueprint/azext_blueprint/tests/latest/input/export_with_artifacts/exported/{self.kwargs['blueprintName']}")
+
+        # recursive function to check for json equality
+        def ordered(obj):
+            if isinstance(obj, dict):
+                return sorted((k, ordered(v)) for k, v in obj.items())
+            if isinstance(obj, list):
+                return sorted(ordered(x) for x in obj)
+            else:
+                return obj
+
+        # file comparison
+        with open(input_blueprint) as input_f:
+            input_blueprint = json.load(input_f)
+            ordered_input_blueprint = ordered(input_blueprint)
+        with open(output_blueprint) as output_f:
+            output_blueprint = json.load(output_f)
+            ordered_output_blueprint = ordered(output_blueprint)
+
+        try:
+            self.assertEqual(ordered_input_blueprint, ordered_output_blueprint)
+        except AssertionError:
+            if output_path.exists() and output_path.is_dir():
+                shutil.rmtree(output_path)
+            raise
+
+        # artifact directory comparison
+        artifacts_cmp = filecmp.dircmp(input_artifact_directory, output_artifact_directory)
+        try:
+            assert len(artifacts_cmp.right_only) == 0 and len(artifacts_cmp.left_only) == 0 and len(artifacts_cmp.funny_files) == 0
+        except AssertionError:
+            if output_path.exists() and output_path.is_dir():
+                shutil.rmtree(output_path)
+            raise
+
+        # artifact file comparison
+        for filename in os.listdir(input_artifact_directory):
+            with open(os.path.join(input_artifact_directory, filename)) as input_f:
+                input_artifact = json.load(input_f)
+                ordered_input_artifact = ordered(input_artifact)
+            with open(os.path.join(output_artifact_directory, filename)) as output_f:
+                output_artifact = json.load(output_f)
+                ordered_output_artifact = ordered(output_artifact)
+            try:
+                self.assertEqual(ordered_input_artifact, ordered_output_artifact)
+            except AssertionError:
+                if output_path.exists() and output_path.is_dir():
+                    shutil.rmtree(output_path)
+                raise
+
+        self.cmd('az blueprint delete '
+                 '--name "{blueprintName}" '
+                 '-y',
+                 checks=[JMESPathCheck('name', self.kwargs.get('blueprintName', ''))])
+
+        if output_path.exists() and output_path.is_dir():
+            shutil.rmtree(output_path)