+++ src/quantum/azext_quantum/vendored_sdks/azure_mgmt_quantum/models/quantum_management_client_enums.py
@@ -43,3 +43,11 @@ class ResourceIdentityType(str, Enum):
 
     system_assigned = "SystemAssigned"
     none = "None"
+
+
+class CreatedByType(str, Enum):
+
+    user = "User"
+    application = "Application"
+    managed_identity = "ManagedIdentity"
+    key = "Key"