+++ src/quantum/azext_quantum/vendored_sdks/azure_mgmt_quantum/models/sku_description.py
@@ -19,8 +19,13 @@ class SkuDescription(Model):
     :type id: str
     :param name: Display name of this sku.
     :type name: str
+    :param version: Display name of this sku.
+    :type version: str
     :param description: Description about this sku.
     :type description: str
+    :param restricted_access_uri: Uri to subscribe to the restricted access
+     sku.
+    :type restricted_access_uri: str
     :param targets: The list of targets available for this sku.
     :type targets: list[str]
     :param quota_dimensions: The list of quota dimensions for this sku.
@@ -32,7 +37,9 @@ class SkuDescription(Model):
     _attribute_map = {
         'id': {'key': 'id', 'type': 'str'},
         'name': {'key': 'name', 'type': 'str'},
+        'version': {'key': 'version', 'type': 'str'},
         'description': {'key': 'description', 'type': 'str'},
+        'restricted_access_uri': {'key': 'restrictedAccessUri', 'type': 'str'},
         'targets': {'key': 'targets', 'type': '[str]'},
         'quota_dimensions': {'key': 'quotaDimensions', 'type': '[QuotaDimension]'},
         'pricing_details': {'key': 'pricingDetails', 'type': '[PricingDetail]'},
@@ -42,7 +49,9 @@ class SkuDescription(Model):
         super(SkuDescription, self).__init__(**kwargs)
         self.id = kwargs.get('id', None)
         self.name = kwargs.get('name', None)
+        self.version = kwargs.get('version', None)
         self.description = kwargs.get('description', None)
+        self.restricted_access_uri = kwargs.get('restricted_access_uri', None)
         self.targets = kwargs.get('targets', None)
         self.quota_dimensions = kwargs.get('quota_dimensions', None)
         self.pricing_details = kwargs.get('pricing_details', None)