+++ src/cloudservice/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.2.0
+++++++
+* Improve user experience.
+
 0.1.0
 ++++++
 * Initial release.